pyproject.toml
[project]
name = "python-tlob-revised"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "ipykernel>=7.0.1",
    "kagglehub>=0.3.13",
    "lightning>=2.5.5",
    "lion-pytorch>=0.2.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "omegaconf>=2.3.0",
    "onnx>=1.19.1",
    "onnxruntime-gpu>=1.23.0",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "pytorch-lightning>=2.5.5",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "torch==2.5.0",
    "torch-ema>=0.3",
    "transformers>=4.57.1",
    "wandb>=0.22.2",
]
